requirements.txt
Flask
requests
python-dotenv
scipy
matplotlib
